PMM-13633 Add hashing even for tokens.
